fix(consensus): retain old epoch information until full transition (#1021)
Retains the epoch state of the outgoing epoch until the node is fully
transitioned to the new (by observing a finalized block in the new
epoch).

Uses the retained outgoing epoch state to start a consensus engine
backing the old epoch on restart.

While working on this patch, it was discovered that the node would also
not start into the incoming epoch: because the boundary block was not
yet finalized and present, the new incoming epoch did not have its
"genesis" and refused to start.

This patch also provides a test to cover this scenario: by looping over
the consensus metrics every 100ms, it attempts to catch a node just as
it processed the boundary-1-th block and restarts it.


Closes #1020

---------

Co-authored-by: Arsenii Kulikov <klkvrr@gmail.com>